requirements.txt
astroid==2.3.3
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
dnspython==1.16.0
Flask==1.1.1
httpie==2.0.0
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.19
protobuf==3.6.1
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.4.4
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.15
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.11.2
